trying to set PDP to load product info
